rag-langraph-bot/requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
langgraph
azure-search-documents
azure-storage-blob
azure-cosmos
azure-identity
tiktoken
pydantic
asyncio
numpy
